Properly quote and emove bashism in libpostal_data.in
The == comparison works in bash, but not sh, also quote the variable
so it doesn't error when the variable is empty.

This would have stopped the senzing model being installed/used on
any host where /bin/sh pointed to a Bourne compatible shell, rather
than bash compatible, and a (harmless, but potentially confusing)
error regarding operators in the case DATAMODEL wasn't set.